stub out API for optional deps
