+++ Cargo.toml
@@ -19,9 +19,9 @@ status = "actively-developed"
 [dependencies]
 log = { version = "0.4", features = ["std"] }
 regex = { version = "1", optional = true }
-termcolor = "1"
-humantime = "1.1"
-atty = "0.2"
+termcolor = { version = "1", optional = true }
+humantime = { version = "1.1", optional = true }
+atty = { version = "0.2", optional = true }
 
 [[test]]
 name = "regexp_filter"
@@ -32,4 +32,4 @@ name = "log-in-log"
 harness = false
 
 [features]
-default = ["regex"]
+default = ["termcolor", "atty", "humantime", "regex"]
