+++ src/fmt/humantime/shim_impl.rs
@@ -1,0 +1,5 @@
+pub(in ::fmt) mod pub_use {
+    /*
+    Timestamps aren't available when we don't have a `humantime` dependency.
+    */
+}
